+++ R/extract.piar_index.R
@@ -3,7 +3,7 @@ dim_indices <- function(x, i) {
     names(x) <- x
   }
   res <- match(x[i], x, incomparables = NA)
-  if (anyNA(res) || length(res) == 0L) {
+  if (length(res) == 0L || anyNA(res)) {
     stop("subscript out of bounds")
   }
   res
@@ -13,17 +13,21 @@ dim_indices <- function(x, i) {
 #'
 #' Methods to extract and replace index values like a matrix.
 #'
-#' The extraction methods treat `x` as a matrix of index values with
+#' The extraction methods treat `x` like a matrix of index values with
 #' (named) rows for each `level` and columns for each `period` in
 #' `x`. Unlike a matrix, dimensions are never dropped as subscripting
 #' `x` always returns an index object. This means that subscripting with a
-#' matrix is not possible, and only a submatrix can be extracted. As `x`
+#' matrix is not possible, and only a "submatrix" can be extracted. As `x`
 #' is not an atomic vector, subscripting with a single index like `x[1]`
 #' extracts all time periods for that level.
 #'
-#' The replacement methods similarly treat `x` as a matrix, and behave the
-#' same as replacing values in a matrix (except that `value` is coerced to
-#' numeric). Note that replacing the values of an index will remove the
+#' The replacement method similarly treat `x` like a matrix. If `value` is
+#' an index object with the same number of time periods as `x[i, j]` and
+#' it inherits from the same class as `x`, then the index values and
+#' percent-change contributions of `x[i, j]` are replaced with those for the
+#' corresponding levels of `value`. If `value` is not an index, then it is
+#' coerced to a numeric vector and behaves the same as replacing values in a 
+#' matrix. Note that replacing the values of an index will remove the
 #' corresponding percent-change contributions (if any).
 #'
 #' Subscripting an aggregate index cannot generally preserve the aggregation
@@ -35,7 +39,7 @@ dim_indices <- function(x, i) {
 #' @param x A price index, as made by, e.g., [elemental_index()].
 #' @param i,j Indices for the levels and time periods of a price index. See
 #' details.
-#' @param value A numeric vector.
+#' @param value A numeric vector or price index. See details.
 #' @param ... Ignored.
 #'
 #' @returns
@@ -89,18 +93,32 @@ dim_indices <- function(x, i) {
   }
 }
 
+# FIXME: Does it make sense to bundle the contributions with the index values
+# as a list with a dim?
 #' @rdname sub-.piar_index
 #' @export
 `[<-.piar_index` <- function(x, i, j, ..., value) {
   levels <- dim_indices(x$levels, i)
   periods <- dim_indices(x$time, j)
-  res <- as.matrix(x)
-  res[levels, periods] <- as.numeric(value)
-  # only loop over periods that have a value replaced
-  for (t in periods) {
-    x$index[[t]][levels] <- res[levels, t]
-    # drop contributions for replaced values
-    x$contrib[[t]][levels] <- list(numeric(0L))
+  if (is_index(value)) {
+    if (length(value$time) != length(periods)) {
+      stop("'x' and 'value' must have the same number of time periods")
+    }
+    if (length(levels) %% length(value$levels) != 0) {
+      stop("number of items to replace is not a multiple of replacement length")
+    }
+    for (t in seq_along(periods)) {
+      x$index[[periods[t]]][levels] <- value$index[[t]]
+      x$contrib[[periods[t]]][levels] <- value$contrib[[t]]
+    }
+  } else {
+    # do all the replacements with a matrix of x so that value is recycled
+    res <- as.matrix(x)
+    res[levels, periods] <- as.numeric(value)
+    for (t in periods) {
+      x$index[[t]][levels] <- res[levels, t]
+      x$contrib[[t]][levels] <- list(numeric(0L))
+    }
   }
   validate_piar_index(x)
 }
@@ -111,5 +129,21 @@ dim_indices <- function(x, i) {
     x$index, x$contrib, x$levels, x$time,
     is_chainable_index(x)
   )
+  NextMethod("[<-")
+}
+
+#' @export
+`[<-.chainable_piar_index` <- function(x, i, j, ..., value) {
+  if (is_index(value) && !is_chainable_index(value)) {
+    stop("'value' must be a period-over-period index")
+  }
+  NextMethod("[<-")
+}
+
+#' @export
+`[<-.direct_piar_index` <- function(x, i, j, ..., value) {
+  if (is_index(value) && !is_direct_index(value)) {
+    stop("'value' must be a fixed-base index")
+  }
   NextMethod("[<-")
 }