+++ R/update.aggregation_structure.R
@@ -14,7 +14,7 @@
 #' otherwise, the default is 1 for an arithmetic index.
 #' @param ... Further arguments passed to or used by methods.
 #'
-#' @return
+#' @returns
 #' A copy of `object` with price-updated weights using the index
 #' values in `index`.
 #'
@@ -43,9 +43,9 @@
 #'   matrix(1:9, 3, dimnames = list(c("111", "112", "121"), NULL))
 #' )
 #'
-#' weights(pias)
+#' weights(pias, ea_only = FALSE)
 #'
-#' weights(update(pias, index))
+#' weights(update(pias, index), ea_only = FALSE)
 #'
 #' @importFrom stats update
 #' @family aggregation structure methods